pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "floodgate"
version = "0.1.0"
description = "Zero-hop MQTT anti-flood service for Meshtastic networks"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.80.0",
    "protobuf>=4.25.0",
    "paho-mqtt>=2.0.0",
    "pyyaml>=6.0",
    "python-json-logger>=2.0",
]

[project.optional-dependencies]
dev = [
    "grpcio-tools>=1.80.0",
    "pytest>=7.0",
]

[project.scripts]
floodgate = "floodgate.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
